feat: altera tipo de dados de preços e totais para BigDecimal nas entidades de movimentação e remoção de código antigo inacabado

+++ src/main/java/sptech/school/dto/itensNaMovimentacao/ItensNaMovimentacaoRequestDto.java
@@ -2,6 +2,8 @@ package sptech.school.dto.itensNaMovimentacao;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ItensNaMovimentacaoRequestDto(
         @NotNull
         Integer movimentacaoEstoqueId,
@@ -11,7 +13,7 @@ public record ItensNaMovimentacaoRequestDto(
 
         Integer qtd,
 
-        Double precoUnitario
+        BigDecimal precoUnitario
 ) {
 }
