+++ src/main/java/sptech/school/dto/movimentacaoEstoque/MovimentacaoEstoqueRequestDto.java
@@ -2,13 +2,14 @@ package sptech.school.dto.movimentacaoEstoque;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record MovimentacaoEstoqueRequestDto(
         @NotNull
         Long usuarioId,
-        Double totalGastoImpostos,
-        Double precoFrete,
+        BigDecimal totalGastoImpostos,
+        BigDecimal precoFrete,
         LocalDate dataEntregaPrevista,
         LocalDate dataEntrega,
         String observacoes,
